current/requirements.txt
fastapi>=0.110.0
uvicorn>=0.22.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pydantic>=2.0
python-dateutil==2.8.2 
openpyxl